value,text,logを分離
