+++ src/value.ts
@@ -1,7 +1,6 @@
 import { Member } from "./member.js";
 import { EventTarget, eventType } from "./event.js";
 import { Field } from "./field.js";
-import { LogLine } from "./logger.js";
 
 /**
  * Valueを指すクラス
@@ -105,69 +104,5 @@ export class Value extends EventTarget<Value> {
    */
   time() {
     return this.data.syncTimeStore.getRecv(this.member_) || new Date(0);
-  }
-}
-export class Text extends EventTarget<Text> {
-  constructor(base: Field, field = "") {
-    super("", base.data, base.member_, field || base.field_);
-    this.eventType_ = eventType.textChange(this);
-  }
-  get member() {
-    return new Member(this);
-  }
-  get name() {
-    return this.field_;
-  }
-  child(field: string): Text {
-    return new Text(this, this.field_ + "." + field);
-  }
-  tryGet() {
-    return this.data.textStore.getRecv(this.member_, this.field_);
-  }
-  get() {
-    const v = this.tryGet();
-    if (v === null) {
-      return "";
-    } else {
-      return v;
-    }
-  }
-  set(data: string | object) {
-    if (this.data.textStore.isSelf(this.member_)) {
-      if (typeof data === "object" && data != null) {
-        for (const [k, v] of Object.entries(data)) {
-          this.child(k).set(v as string | object);
-        }
-      } else {
-        this.data.textStore.setSend(this.field_, String(data));
-        this.triggerEvent(this);
-      }
-    } else {
-      throw new Error("Cannot set data to member other than self");
-    }
-  }
-  time() {
-    return this.data.syncTimeStore.getRecv(this.member_) || new Date(0);
-  }
-}
-
-export class Log extends EventTarget<Log> {
-  constructor(base: Field) {
-    super("", base.data, base.member_, "");
-    this.eventType_ = eventType.logAppend(this);
-  }
-  get member() {
-    return new Member(this);
-  }
-  tryGet() {
-    return this.data.logStore.getRecv(this.member_);
-  }
-  get() {
-    const v = this.tryGet();
-    if (v === null) {
-      return [];
-    } else {
-      return v;
-    }
   }
 }
